backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy>=2.0.36
alembic==1.13.1
asyncpg==0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
mcp>=1.0.0

python-multipart==0.0.9
email-validator==2.1.1
cryptography==42.0.7
openai>=1.0.0
pgvector>=0.2.0
greenlet
